finish migration to intake subsystem
