+++ src/main/java/frc/robot/subsystems/IntakeSubsystem.java
@@ -2,12 +2,14 @@ package frc.robot.subsystems;
 import com.revrobotics.CANSparkMax;
 import com.revrobotics.CANSparkMaxLowLevel.MotorType;
 import frc.robot.Constants.ArmConstants;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 public class IntakeSubsystem {
   private CANSparkMax leftIntakeMotor, rightIntakeMotor, topIntakeMotor;
-  private double intakeGoalPower;
+  private double intakePower;
   private double intakeOff = 0;
   private double intakeOn = ArmConstants.intakeVoltage;
+  private final double HARD_STOP = 25.0;
 
   public IntakeSubsystem() {
     leftIntakeMotor = new CANSparkMax(ArmConstants.leftIntakeMotorID, MotorType.kBrushless);
@@ -19,9 +21,27 @@ public class IntakeSubsystem {
     topIntakeMotor.setSmartCurrentLimit(80);
   }
   public void intakeOn () {
-    intakeGoalPower = intakeOn;
+    intakePower = intakeOn;
   }
   public void intakeOff () {
-    intakeGoalPower = intakeOff;
+    intakePower = intakeOff;
+  }
+  public void setIntakeMotorPower(double percent) {
+    rightIntakeMotor.set(percent); // left motor follows?
+    topIntakeMotor.set(percent);
+  }
+  public double getIntakeMotorAmps() {
+    return rightIntakeMotor.getOutputCurrent();
+  }
+  private void checkIntakeAmps() {
+    if(getIntakeMotorAmps() > HARD_STOP) {
+      setIntakeMotorPower(0);
+    }
+  }
+  @Override
+  public void periodic() {
+    setIntakeMotorPower(intakeGoalPower);
+    intakeOn = SmartDashboard.getNumber("intake power", 0);
+    checkIntakeAmps();
   }
 }
